downstream-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-dotenv==1.0.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2